refactor: 3rdparty/uuid
